relocate TestcaseParser
